pricelists

+++ app/tenants/organization.js
@@ -1123,33 +1123,41 @@ docReady(function () {
     var table = $("#table_pricelists_list").DataTable();
     var tr = $(this).closest("tr");
     var rowData = table.row(tr).data();
-    console.log(rowData)
-    var endpoint = InvokeURL + "price-lists/" + rowData.uuid;
-    $.ajax({
-      type: "DELETE",
-      url: endpoint,
-      beforeSend: function () {
-        $("#waitingdots").show();
-      },
-      complete: function () {
-        $("#waitingdots").hide();
-      },
-      headers: {
-        Authorization: orgToken
-      },
-      success: function () {
-        console.log("Rekord został pomyślnie usunięty.");
-        $("#waitingdots").show(1).delay(150).hide(1);
-        table.row($(this).parents("tr")).remove().draw();    
-      },
-      error: function (xhr, status, error) {
-        console.error("Błąd usuwania rekordu:", error);
-
+  
+    if (rowData && rowData.uuid) {
+      // Wyświetl potwierdzenie usuwania
+      var confirmDelete = confirm("Czy na pewno chcesz usunąć ten rekord?");
+      
+      if (confirmDelete) {
+        var endpoint = InvokeURL + "price-lists/" + rowData.uuid;
+        
+        $.ajax({
+          type: "DELETE",
+          url: endpoint,
+          beforeSend: function () {
+            $("#waitingdots").show();
+          },
+          complete: function () {
+            $("#waitingdots").hide();
+          },
+          headers: {
+            Authorization: orgToken
+          },
+          success: function () {
+            console.log("Rekord został pomyślnie usunięty.");
+            $("#waitingdots").show(1).delay(150).hide(1);
+            table.row($(this).parents("tr")).remove().draw();    
+          },
+          error: function (xhr, status, error) {
+            console.error("Błąd usuwania rekordu:", error);
+          }
+        });
       }
-    });
-
-    
+    } else {
+      console.error("Brak UUID w danych rekordu.");
+    }
   });
+  
 
   makeWebflowFormAjaxNewWh = function (forms, successCallback, errorCallback) {
     forms.each(function () {
